--- data.html
@@ -1,237 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Database View</title>
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" />
-    <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script>
-    <style>
-        body, html {
-            margin: 0;
-            padding: 0;
-            font-family: Arial, sans-serif;
-        }
-
-        .menu-bar {
-            background-color: #333;
-            color: white;
-            display: flex;
-            align-items: center;
-            padding: 0 15px;
-            height: 60px;
-            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-            position: relative;
-            z-index: 1000;
-        }
-
-        .menu-bar .logo {
-            height: 50px;
-            width: auto;
-            object-fit: contain;
-            margin-right: 15px;
-        }
-
-        .menu-bar .menu {
-            display: flex;
-            align-items: center;
-            list-style: none;
-            margin: 0;
-            padding: 0;
-        }
-
-        .menu-bar .menu-item {
-            position: relative;
-            margin: 0 15px;
-        }
-
-        .menu-bar .menu-item a {
-            text-decoration: none;
-            color: white;
-            font-size: 18px;
-            padding: 5px 10px;
-            transition: color 0.3s;
-        }
-
-        .menu-bar .menu-item a:hover {
-            color: #ff9800;
-        }
-
-        .dropdown {
-            display: none;
-            position: absolute;
-            top: 100%; 
-            left: 0;
-            background-color: #444;
-            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-            border-radius: 4px;
-            overflow: hidden;
-            z-index: 1001;
-        }
-
-        .dropdown a {
-            display: block;
-            padding: 10px 15px;
-            color: white;
-            text-decoration: none;
-            font-size: 16px;
-        }
-
-        .dropdown a:hover {
-            background-color: #555;
-            color: #ff9800;
-        }
-
-        .menu-item:hover .dropdown {
-            display: block;
-        }
-
-        .content {
-            padding: 20px;
-            max-width: 1200px;
-            margin: auto;
-        }
-
-        table {
-            width: 100%;
-            border-collapse: collapse;
-            margin: 20px 0;
-        }
-
-        th, td {
-            border: 1px solid #ddd;
-            text-align: left;
-            padding: 8px;
-        }
-
-        th {
-            background-color: #f4f4f4;
-        }
-
-        .download-btn {
-            display: inline-block;
-            padding: 10px 15px;
-            background-color: #007BFF;
-            color: white;
-            text-decoration: none;
-            border-radius: 5px;
-            font-size: 16px;
-        }
-
-        .download-btn:hover {
-            background-color: #0056b3;
-        }
-
-        .filter-container {
-            margin: 20px 0;
-        }
-
-        .filter-container select {
-            padding: 8px;
-            font-size: 16px;
-        }
-    </style>
-</head>
-<body>
-    <div class="menu-bar">
-        <img src="team-logo.png" alt="Team Logo" class="logo">
-        <ul class="menu">
-            <li class="menu-item">
-                <a href="index.html">Tracking</a>
-                <div class="dropdown">
-                    <a href="index.html">Active Wildfires</a>
-                    <a href="past.html">Past Wildfires</a>
-                </div>
-            </li>
-            <li class="menu-item"><a href="data.html">Data</a></li>
-            <li class="menu-item"><a href="about.html">About</a></li>
-            <li class="menu-item"><a href="contact.html">Contact Us</a></li>
-        </ul>
-    </div>
-
-    <div class="content">
-        <h1>Database View</h1>
-        <p>Below is the current data stored in the database:</p>
-        <div class="filter-container">
-            <label for="fire-filter">Filter by Fire Name:</label>
-            <select id="fire-filter">
-                <option value="">All Fires</option>
-            </select>
-        </div>
-        <a href="/download_csv" class="download-btn">Download as CSV</a>
-        <div id="table-container">
-        </div>
-    </div>
-
-    <script>
-        async function loadTableData(fireName = "") {
-            try {
-                const response = await fetch('/get_database');
-                if (!response.ok) throw new Error('Failed to fetch database data');
-
-                const data = await response.json();
-                const tableContainer = document.getElementById('table-container');
-                const fireFilter = document.getElementById('fire-filter');
-
-                const fireNames = [...new Set(data.map(row => row.name))];
-                if (fireFilter.options.length === 1) { 
-                    fireNames.forEach(name => {
-                        const option = document.createElement('option');
-                        option.value = name;
-                        option.textContent = name;
-                        fireFilter.appendChild(option);
-                    });
-                }
-
-                const filteredData = fireName ? data.filter(row => row.name === fireName) : data;
-
-                if (filteredData.length === 0) {
-                    tableContainer.innerHTML = '<p>No data available for the selected fire.</p>';
-                    return;
-                }
-
-                const table = document.createElement('table');
-                const thead = document.createElement('thead');
-                const tbody = document.createElement('tbody');
-
-                const headers = Object.keys(filteredData[0]);
-                const headerRow = document.createElement('tr');
-                headers.forEach(header => {
-                    const th = document.createElement('th');
-                    th.textContent = header;
-                    headerRow.appendChild(th);
-                });
-                thead.appendChild(headerRow);
-
-                filteredData.forEach(row => {
-                    const tr = document.createElement('tr');
-                    headers.forEach(header => {
-                        const td = document.createElement('td');
-                        td.textContent = row[header];
-                        tr.appendChild(td);
-                    });
-                    tbody.appendChild(tr);
-                });
-
-                table.appendChild(thead);
-                table.appendChild(tbody);
-                tableContainer.innerHTML = '';
-                tableContainer.appendChild(table);
-            } catch (error) {
-                console.error(error);
-                document.getElementById('table-container').innerHTML = '<p>Failed to load database data.</p>';
-            }
-        }
-
-        document.addEventListener('DOMContentLoaded', () => {
-            loadTableData();
-
-            const fireFilter = document.getElementById('fire-filter');
-            fireFilter.addEventListener('change', () => {
-                const selectedFire = fireFilter.value;
-                loadTableData(selectedFire);
-            });
-        });
-    </script>
-</body>
-</html>
