+++ gcs/packet_receiver.py
@@ -2,8 +2,32 @@ import struct
 import zlib
 import sys
 import serial
+import json
+import requests
 from packet_class._v2.packet import Packet
-PACKET_SIZE = 24 # ADJUST?
+
+PACKET_SIZE = 24  # ADJUST?
+
+# Possible function to forward packets to server alongside Robb's current implementation in receive_and_decode_packets
+def send_packet_to_server(packet):
+    """Sends the decoded packet to the server."""
+    server_url = "http://localhost:8000/add_packet"  # Current Server Location
+    try:
+        packet_data = {
+            "pac_id": packet.pac_id,
+            "gps_data": packet.gps_data,
+            "alt": packet.alt,
+            "high_temp": packet.high_temp,
+            "low_temp": packet.low_temp,
+        }
+
+        response = requests.post(server_url, json=packet_data)
+        if response.status_code == 200:
+            print("Packet successfully sent to server.")
+        else:
+            print(f"Failed to send packet to server. Status code: {response.status_code}, Response: {response.text}")
+    except requests.RequestException as e:
+        print(f"Error connecting to the server: {e}")
 
 def receive_and_decode_packets():
     # Open the serial port connected to the RF module
@@ -46,8 +70,9 @@ def receive_and_decode_packets():
                 low_temp=low_temp
             )
 
-            # Print the decoded packet
+            # Print the decoded packet and send to the server
             print(packet)
+            send_packet_to_server(packet)
 
         except struct.error as e:
             print(f"Error decoding packet: {e}")
