+++ gcs/scorch/assets/team-logo.png
