+++ gcs/scorch/past.html
@@ -1,0 +1,194 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Past Wildfires</title>
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" />
+    <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script>
+    <style>
+        body, html {
+            margin: 0;
+            padding: 0;
+            font-family: Arial, sans-serif;
+        }
+
+        .menu-bar {
+            background-color: #333;
+            color: white;
+            display: flex;
+            align-items: center;
+            padding: 0 15px;
+            height: 60px;
+            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+            position: relative;
+            z-index: 1000;
+        }
+
+        .menu-bar .logo {
+            height: 50px;
+            width: auto;
+            object-fit: contain;
+            margin-right: 15px;
+        }
+
+        .menu-bar .menu {
+            display: flex;
+            align-items: center;
+            list-style: none;
+            margin: 0;
+            padding: 0;
+        }
+
+        .menu-bar .menu-item {
+            position: relative;
+            margin: 0 15px;
+        }
+
+        .menu-bar .menu-item a {
+            text-decoration: none;
+            color: white;
+            font-size: 18px;
+            padding: 5px 10px;
+            transition: color 0.3s;
+        }
+
+        .menu-bar .menu-item a:hover {
+            color: #ff9800;
+        }
+
+        .menu-bar .dropdown {
+            display: none; 
+            position: absolute;
+            top: 100%; 
+            left: 0;
+            background-color: #444;
+            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+            border-radius: 4px;
+            overflow: hidden;
+            z-index: 1001;
+        }
+
+        .menu-bar .dropdown a {
+            display: block;
+            padding: 10px 15px;
+            color: white;
+            text-decoration: none;
+            font-size: 16px;
+            transition: background-color 0.3s, color 0.3s;
+        }
+
+        .menu-bar .dropdown a:hover {
+            background-color: #555;
+            color: #ff9800;
+        }
+
+        .menu-bar .menu-item:hover .dropdown {
+            display: block;
+        }
+
+        .content {
+            padding: 20px;
+            text-align: center;
+        }
+
+        .search-bar {
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            gap: 10px;
+            margin-bottom: 20px;
+        }
+
+        #map {
+            height: 600px;
+            width: 100%;
+            margin-top: 20px;
+        }
+    </style>
+</head>
+<body>
+    <div class="menu-bar">
+        <img src="assets\team-logo.png" alt="Team Logo" class="logo">
+        <ul class="menu">
+            <li class="menu-item">
+                <a href="index.html">Tracking</a>
+                <div class="dropdown">
+                    <a href="index.html">Active Wildfires</a>
+                    <a href="past.html">Past Wildfires</a>
+                </div>
+            </li>
+            <li class="menu-item"><a href="data.html">Data</a></li>
+            <li class="menu-item"><a href="about.html">About</a></li>
+            <li class="menu-item"><a href="contact.html">Contact Us</a></li>
+        </ul>
+    </div>
+
+    <div class="content">
+        <h1>Past Wildfires</h1>
+        <p>Previous wildfires that we have tracked:</p>
+        
+        <div class="search-bar">
+            <input type="text" id="search-location" placeholder="Enter location (e.g., city or ZIP)">
+            <span>or</span>
+            <select id="wildfire-select">
+                <option value="">Select Fire Name</option>
+            </select>
+            <button onclick="searchMap()">Search</button>
+        </div>
+
+        <div id="map"></div>
+    </div>
+
+    <script src="javascript/map_script.js"></script>
+    <script>
+        document.addEventListener("DOMContentLoaded", () => {
+            loadMap('past');
+            loadFireNames();
+        });
+
+        async function loadFireNames() {
+            const response = await fetch('/wildfire_markers?filter=past');
+            const fireData = await response.json();
+
+            const fireNameSelect = document.getElementById('wildfire-select');
+            fireData.forEach(fire => {
+                const option = document.createElement('option');
+                option.value = fire.name;
+                option.textContent = fire.name;
+                fireNameSelect.appendChild(option);
+            });
+        }
+
+        async function searchMap() {
+            const location = document.getElementById('search-location').value;
+            const fireName = document.getElementById('wildfire-select').value;
+
+            if (location) {
+                const response = await fetch(`https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(location)}`);
+                const locationData = await response.json();
+
+                if (locationData.length > 0) {
+                    const { lat, lon } = locationData[0];
+                    window.map.setView([lat, lon], 12);
+                } else {
+                    alert('Location not found');
+                }
+            } else if (fireName) {
+                const response = await fetch(`/wildfire_markers?filter=past`);
+                const fireData = await response.json();
+                const selectedFire = fireData.find(fire => fire.name === fireName);
+
+                if (selectedFire) {
+                    window.map.setView([selectedFire.latitude, selectedFire.longitude], 12);
+                } else {
+                    alert('Fire not found');
+                }
+            } else {
+                alert('Please enter a location or select a fire name');
+            }
+        }
+    </script>
+
+</body>
+</html>
