+++ gcs/scorch/wildfire-icon.webp
